+++ python/common/src/piscsi/piscsi_cmds.py
@@ -2,10 +2,21 @@
 Module for commands sent to the PiSCSI backend service.
 """
 
+import logging
+from pathlib import PurePath, Path
+from functools import lru_cache
+
 import piscsi_interface_pb2 as proto
 from piscsi.return_codes import ReturnCodes
 from piscsi.socket_cmds import SocketCmds
-import logging
+
+from piscsi.common_settings import (
+    CFG_DIR,
+    PROPERTIES_SUFFIX,
+    ARCHIVE_FILE_SUFFIXES,
+)
+
+from util import unarchiver
 
 
 class PiscsiCmds:
@@ -23,6 +34,79 @@ class PiscsiCmds:
             logging.debug(self.format_pb_command(command))
 
         return self.sock_cmd.send_pb_command(command.SerializeToString())
+
+    def list_images(self):
+        """
+        Sends a IMAGE_FILES_INFO command to the server
+        Returns a (dict) with (bool) status, (str) msg, and (list) of (dict)s files
+        """
+        from piscsi.file_cmds import FileCmds
+
+        self.file_cmd = FileCmds(piscsi=self)
+
+        command = proto.PbCommand()
+        command.operation = proto.PbOperation.DEFAULT_IMAGE_FILES_INFO
+        command.params["token"] = self.token
+        command.params["locale"] = self.locale
+
+        data = self.send_pb_command(command)
+        result = proto.PbResult()
+        result.ParseFromString(data)
+
+        server_info = self.get_server_info()
+        files = []
+        for file in result.image_files_info.image_files:
+            prop_file_path = Path(CFG_DIR) / f"{file.name}.{PROPERTIES_SUFFIX}"
+            # Add properties meta data for the image, if matching prop file is found
+            if prop_file_path.exists():
+                process = self.file_cmd.read_drive_properties(prop_file_path)
+                prop = process["conf"]
+            else:
+                prop = False
+
+            archive_contents = []
+            if PurePath(file.name).suffix.lower()[1:] in ARCHIVE_FILE_SUFFIXES:
+                try:
+                    archive_info = self._get_archive_info(
+                        f"{server_info['image_dir']}/{file.name}",
+                        _cache_extra_key=file.size,
+                    )
+
+                    properties_files = [
+                        x["path"]
+                        for x in archive_info["members"]
+                        if x["path"].endswith(PROPERTIES_SUFFIX)
+                    ]
+
+                    for member in archive_info["members"]:
+                        if member["is_dir"] or member["is_resource_fork"]:
+                            continue
+
+                        if PurePath(member["path"]).suffix.lower()[1:] == PROPERTIES_SUFFIX:
+                            member["is_properties_file"] = True
+                        elif f"{member['path']}.{PROPERTIES_SUFFIX}" in properties_files:
+                            member[
+                                "related_properties_file"
+                            ] = f"{member['path']}.{PROPERTIES_SUFFIX}"
+
+                        archive_contents.append(member)
+                except (unarchiver.LsarCommandError, unarchiver.LsarOutputError):
+                    pass
+
+            size_mb = "{:,.1f}".format(file.size / 1024 / 1024)
+            dtype = proto.PbDeviceType.Name(file.type)
+            files.append(
+                {
+                    "name": file.name,
+                    "size": file.size,
+                    "size_mb": size_mb,
+                    "detected_type": dtype,
+                    "prop": prop,
+                    "archive_contents": archive_contents,
+                }
+            )
+
+        return {"status": result.status, "msg": result.msg, "files": files}
 
     def get_server_info(self):
         """
@@ -521,3 +605,15 @@ class PiscsiCmds:
             message += f", device: {formatted_device}"
 
         return message
+
+    # noinspection PyMethodMayBeStatic
+    @lru_cache(maxsize=32)
+    def _get_archive_info(self, file_path, **kwargs):
+        """
+        Cached wrapper method to improve performance, e.g. on index screen
+        """
+        try:
+            return unarchiver.inspect_archive(file_path)
+        except (unarchiver.LsarCommandError, unarchiver.LsarOutputError) as error:
+            logging.error(str(error))
+            raise
